refactor: 로그 메시지 렌더링 기능 이동
- renderLogMessage 함수를 view.js로 이동
